renaming user model file from "user" to "models" in imports
