+++ src/api/routes/checkins/checkins.py
@@ -3,7 +3,7 @@ from flask_jwt_extended import jwt_required, get_jwt_identity
 
 from src.database.db import db
 from src.database.checkins.service import CheckinService, CheckinTargetNotFound, CheckinValidationError
-from src.database.users.user import User
+from src.database.users.models import User
 from src.utils.timezone import parse_dt, to_utc, from_utc
